# docs-site/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Greenlight — Multi-Platform Prompt & Channel Gateway",
    "version": "0.1.0",
    "description": "Self-hosted HTTP gateway for human-in-the-loop AI agents across Telegram, Slack, Teams, Discord, Google Chat, WhatsApp, and Messenger.\n\n**Path verbs:** `POST …/new` creates a durable resource (channel or prompt). `POST …/send` delivers chat text on a MESSAGE channel only.",
    "license": {
      "name": "BUSL-1.1",
      "url": "https://spdx.org/licenses/BUSL-1.1.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8100",
      "description": "Local development"
    },
    {
      "url": "https://{host}",
      "description": "Self-hosted production (PUBLIC_WEBHOOK_URL origin)",
      "variables": {
        "host": {
          "default": "api.example.com",
          "description": "Your Greenlight host (no scheme)"
        }
      }
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Database-backed API key with scoped permissions. Bootstrap from GREENLIGHT_API_KEY env on first boot when api_keys is empty."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      },
      "Platform": {
        "type": "string",
        "enum": [
          "telegram",
          "slack",
          "teams",
          "discord",
          "gchat",
          "whatsapp",
          "messenger"
        ]
      },
...
      "CreatePromptRequest": {
        "type": "object",
        "required": [
          "text"
        ],
        "properties": {
          "channel_id": {
            "type": "string",
            "nullable": true,
            "description": "Target PROMPT channel. Optional if DEFAULT_PROMPT_CHANNEL_ID is set."
          },
          "text": {
            "type": "string",
            "maxLength": 4096
          },
          "options": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 64
            },
            "maxItems": 10,
            "nullable": true,
            "description": "Button labels. Max 3 on WhatsApp/Messenger."
          },
          "allow_text": {
            "type": "boolean",
            "default": false,
            "description": "Allow free-text replies matching ID:#N patterns."
          },
          "callback_url": {
            "type": "string",
            "nullable": true,
            "description": "HTTPS URL for signed answer callbacks."
          },
          "correlation_id": {
            "type": "string",
            "maxLength": 255,
            "nullable": true
          },
          "ttl_sec": {
            "type": "integer",
            "minimum": 0,
            "maximum": 604800,
            "nullable": true,
            "description": "Default 3600."
          },
          "media_url": {
            "type": "string",
            "nullable": true
          },
          "media_path": {
            "type": "string",
            "nullable": true,
            "description": "Path under MEDIA_ALLOWED_DIR."
          }
        }
      },
      "CreatePromptResponse": {
        "type": "object",
        "required": [
          "prompt_id",
          "channel_id",
          "message_id"
        ],
        "properties": {
          "prompt_id": {
            "type": "string",
            "example": "#123",
            "description": "Prompt id including leading #."
          },
          "channel_id": {
            "type": "string"
          },
          "message_id": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ],
            "description": "Platform message identifier."
          }
        }
      },
      "Prompt": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "#123"
          },
          "prompt_num": {
            "type": "integer"
          },
          "chat_id": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "media_url": {
            "type": "string",
            "nullable": true
          },
          "options": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allow_text": {
            "type": "boolean"
          },
          "callback_url": {
            "type": "string",
            "nullable": true
          },
          "correlation_id": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "enum": [
              "pending",
              "answered",
              "expired"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "answered_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "answered_by_id": {
            "type": "string",
            "nullable": true
          },
          "answered_by_username": {
            "type": "string",
            "nullable": true
          },
          "answer": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "SendMessageRequest": {
        "type": "object",
        "required": [
          "channel_id",
          "text"
        ],
        "properties": {
          "channel_id": {
            "type": "string",
            "minLength": 1
          },
          "text": {
            "type": "string",
            "minLength": 1
          }
        }
      },
...
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "channel_id": {
            "type": "string"
          },
          "direction": {
            "type": "string",
            "enum": [
              "inbound",
              "outbound"
            ]
          },
          "text": {
            "type": "string"
          },
          "platform": {
            "$ref": "#/components/schemas/Platform"
          },
          "from_user": {
            "type": "string",
            "nullable": true
          },
          "api_key_id": {
            "type": "string",
            "nullable": true,
            "description": "Outbound API key id"
          },
          "platform_message_id": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/v1/prompts/new": {
      "post": {
        "tags": [
          "Prompts"
        ],
        "summary": "Create a prompt",
        "description": "Creates a durable prompt resource and posts it to the PROMPT channel. Accepts `application/json` or `multipart/form-data` (for file upload). Use `/send` only for MESSAGE channel chat text.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePromptRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "text"
                ],
                "properties": {
                  "text": {
                    "type": "string"
                  },
                  "channel_id": {
                    "type": "string"
                  },
                  "options": {
                    "type": "string",
                    "description": "JSON array of option strings"
                  },
                  "allow_text": {
                    "type": "string"
                  },
                  "callback_url": {
                    "type": "string"
                  },
                  "correlation_id": {
                    "type": "string"
                  },
                  "ttl_sec": {
                    "type": "string"
                  },
                  "media_url": {
                    "type": "string"
                  },
                  "media_path": {
                    "type": "string"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prompt created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePromptResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation or business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Content-Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/prompts": {
      "get": {
        "tags": [
          "Prompts"
        ],
        "summary": "List prompts",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "answered",
                "expired",
                "all"
              ],
              "default": "pending"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Prompt list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Prompt"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/prompts/{id}": {
      "get": {
        "tags": [
          "Prompts"
        ],
        "summary": "Get prompt by id",
        "description": "Encode `#` as `%23` (e.g. `%23123` for `#123`).",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "%23123"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Prompt found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Prompt"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/messages": {
      "get": {
        "tags": [
          "Messages"
        ],
        "summary": "List message history",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "direction",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "inbound",
                "outbound",
                "all"
              ],
              "default": "all"
            }
          },
          {
            "name": "channel_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/messages/send": {
      "post": {
        "tags": [
          "Messages"
        ],
        "summary": "Send chat text on a MESSAGE channel",
        "description": "Delivers plain text in an existing MESSAGE channel thread. Use `/new` to create channels or prompts.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sent",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "sent"
                    },
                    "message_id": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Not a MESSAGE channel or send error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Channel not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/messages/{id}": {
      "get": {
        "tags": [
          "Messages"
        ],
        "summary": "Get message by id",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}